device-posture/device-posture-event-listener.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script>
'use strict';

// REFERENCE https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/web_tests/http/tests/inspector-protocol/device-posture/device-posture-event-listener.js
promise_test(async (t) => {
  assert_equals(navigator.devicePosture.type, 'continuous');

  const promise = new Promise(resolve => {
    navigator.devicePosture.addEventListener(
      'change',
      () => { resolve(navigator.devicePosture.type); },
      { once: true }
    );
  });
  await test_driver.set_device_posture('folded');
  assert_equals(await promise, 'folded')
}, 'Device posture test: device-posture-event-listener.js.');
</script>
